requirements.txt
absl-py==2.3.1
astunparse==1.6.3
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
contourpy==1.3.0
cycler==0.12.1
flatbuffers==25.2.10
fonttools==4.59.0
gast==0.6.0
google-auth==2.40.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.73.1
h5py==3.14.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
keras==2.14.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.8.2
MarkupSafe==3.0.2
matplotlib==3.9.4
ml-dtypes==0.2.0
numpy==2.0.2
oauthlib==3.3.1
opencv-python==4.12.0.88
opt_einsum==3.4.0
packaging==25.0
pillow==11.3.0
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.14.0
termcolor==3.1.0
tqdm==4.67.1
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.14.1
zipp==3.23.0
